backend/pyproject.toml
[tool.poetry]
name = "ai-meeting-kb-backend"
version = "1.0.0"
description = "AI Meeting Knowledge SaaS Backend"
authors = ["Team"]
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pgvector = "^0.3.0"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.0.0"
python-telegram-bot = {extras = ["all"], version = "^21.0"}
litellm = "^1.40.0"
cryptography = "^42.0.0"
httpx = "^0.27.0"
google-auth = "^2.29.0"
google-auth-oauthlib = "^1.2.0"
google-api-python-client = "^2.131.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
